packages/ai-processor/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and ML libraries
openai==1.3.7
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2
scikit-image==0.22.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2

# Computer vision and image processing
ultralytics==8.0.206
easyocr==1.7.0
pytesseract==0.3.10

# Audio processing
elevenlabs==0.2.26
pydub==0.25.1

# Data processing
pandas==2.1.3
redis==5.0.1
celery==5.3.4

# HTTP and networking
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0